\name{SimilarityMeasures-package}
\alias{SimilarityMeasures-package}
\alias{SimilarityMeasures}
\docType{package}
\title{
Implements Several Similarity Measures and Useful Functions
}
\description{
This package contains functions to run and assist four different similarity measures. The similarity measures included are: longest common subsequence (LCSS), Frechet distance, edit distance and dynamic time warping (DTW). Each of these similarity measures can be calculated from two n-dimensional trajectories, both in matrix form.
}
\details{
\tabular{ll}{
Package: \tab SimilarityMeasures\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2014-02-06\cr
License: \tab GPL-3\cr
}
}
\author{
Kevin Toohey

Maintainer: Kevin Toohey <kevintoohey@live.com>
}
\references{
Alt, H. and Godau, M. (1995) Computing the Frechet distance between two polygonal curves. \emph{International Journal of Computational Geometry & Applications}, \bold{5(01n02)}, 75--91.

Berndt, D. and Clifford, J. (1994) Using Dynamic Time Warping to Find Patterns in Time Series. Paper presented at the KDD workshop.

Chen, L., Ozsu, M. T. and Oria, V. (2005) Robust and fast similarity search for moving object trajectories. Paper presented at the Proceedings of the 2005 ACM SIGMOD international conference on Management of data, Baltimore, Maryland.

Vlachos, M., Kollios, G. and Gunopulos, D. (2002) Discovering similar multidimensional trajectories. Paper presented at the Data Engineering, 2002. Proceedings. 18th International Conference on.
}
\keyword{ package }
